TextureAtlasTree 레거시 폴더로 보내버리고, stb_rect_pack 이용하는 TextureAtlas 클래스 새로 만듬.

+++ Source/TGON/String/StringHash.h
@@ -34,7 +34,7 @@ public:
     {
     }
 
-    constexpr BaseBasicStringHash(uint32_t hashCode) noexcept :
+    constexpr BaseBasicStringHash(size_t hashCode) noexcept :
         m_hashCode(hashCode)
     {
     }
